requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopy==2.2.0
google-auth==2.5.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2
